update admin api

+++ src/main/java/landvibe/test/service/AdminService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -14,18 +15,30 @@ public class AdminService {
     private final MemberRepository memberRepository;
 
     public List<Member> findMembersApproveFalse() {
-        List<Member> members = memberRepository.findByApproveFalse();
-        return members;
+        Optional<List<Member>> byApproveFalse = memberRepository.findByApproveFalse();
+        if(byApproveFalse.isEmpty()){
+            return null;
+        }
+        return byApproveFalse.get();
     }
 
     @Transactional
-    public void approveMember(Long memberId) {
-        Member member = memberRepository.findById(memberId).get();
-        member.setApprove(true);
+    public void approveMember(Long memberId) throws Exception {
+        Optional<Member> byId = memberRepository.findById(memberId);
+        if(byId.isEmpty()){
+            // 존재하지 않는 회원
+            throw new Exception();
+        }
+        byId.get().setApprove(true);
     }
 
     @Transactional
-    public void refuseMember(Long memberId) {
+    public void refuseMember(Long memberId) throws Exception {
+        Optional<Member> byId = memberRepository.findById(memberId);
+        if(byId.isEmpty()){
+            // 존재하지 않는 회원
+            throw new Exception();
+        }
         memberRepository.deleteById(memberId);
     }
 }
